path: split repository-specific functions
Make path.[ch] contain pure utility functions that do not know about
public-facing libgit2 objects, including `git_repository`.
