+++ src/tree.c
@@ -14,6 +14,7 @@
 #include "tree-cache.h"
 #include "index.h"
 #include "userbuf.h"
+#include "repo_path.h"
 
 #define DEFAULT_TREE_SIZE 16
 #define MAX_FILEMODE_BYTES 6
